+++ tests/conftest.py
@@ -201,39 +201,3 @@ def dataset(request, paths, engine):
         kwargs["names"] = allcols_csv
 
     return Dataset(paths, part_mem_fraction=gpu_memory_frac, cpu=cpu, **kwargs)
-
-
-@contextlib.contextmanager
-def run_triton_server(modelpath):
-    cmdline = [
-        TRITON_SERVER_PATH,
-        "--model-repository",
-        modelpath,
-        "--backend-config=tensorflow,version=2",
-    ]
-    env = os.environ.copy()
-    env["CUDA_VISIBLE_DEVICES"] = "0"
-    with subprocess.Popen(cmdline, env=env) as process:
-        try:
-            with grpcclient.InferenceServerClient("localhost:8001") as client:
-                # wait until server is ready
-                for _ in range(60):
-                    if process.poll() is not None:
-                        retcode = process.returncode
-                        raise RuntimeError(f"Tritonserver failed to start (ret={retcode})")
-
-                    try:
-                        ready = client.is_server_ready()
-                    except tritonclient.utils.InferenceServerException:
-                        ready = False
-
-                    if ready:
-                        yield client
-                        return
-
-                    time.sleep(1)
-
-                raise RuntimeError("Timed out waiting for tritonserver to become ready")
-        finally:
-            # signal triton to shutdown
-            process.send_signal(signal.SIGINT)
